第一次（2.16， 2.17， 2.20）及第二次作业（2.18，2.19，2.23，2.24）全部重新提交

+++ HWK2.20/main.cpp
@@ -1,14 +1,15 @@
+//homework2.19----Han_Wenkang
 #include <iostream>
 
 using namespace std;
 
 int main()
 {
-    int r=0;
-    int d=0;
-    int C=0;
-    int S=0;
-    int pi=3.14159;
+    double r=0;
+    double d=0;
+    double C=0;
+    double S=0;
+    double pi=3.14159;
     cout<<"Enter the r of circle: ";
     cin>>r;
     d=2*r,C=pi*d,S=pi*r*r;
